Remove "term" from subtraction
In "a - b", I don't want "b" to be called a "term". The word "term" is reserved for parts that are added together, so in "a - b", a and -b are terms, but not b (under that meaning of "term"). Before this, that cell was left blank, but now I added the actual word for a number being subtracted: subtrahend.