romove phanotm = signs
I polled 7 instructors at SY, including myself, and 5 preferred
``
1 + 2
= 3
``
compared to
``
  1+2
= 3
``
The other two had no preference. so I think the default style should be to not indent the first line. (When possible, there is still an overall preference for the first line to have both sides of an equals sign.)